refactor

+++ Server/containers/OutgoingMessage.cpp
@@ -5,7 +5,7 @@
 #include <cstring>
 #include "OutgoingMessage.h"
 
-OutgoingMessage::OutgoingMessage(int fd, const char *payload, int size): fd(fd), size(size) {
+OutgoingMessage::OutgoingMessage(int fd, const char *payload, unsigned long size): fd(fd), size(size) {
     this->payload = new char[size + 1];
     strcpy(this->payload, payload);
 }
